--- src/main/java/org/ezcode/codetest/common/annotation/Auth.java
@@ -1,12 +1,0 @@
-package org.ezcode.codetest.common.annotation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.PARAMETER, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Auth {
-}
